requirements.txt
requests
fastapi
uvicorn
python-dateutil
